Fix namespacing for PyPi installs
From
https://pythonhosted.org/setuptools/setuptools.html#namespace-packages

"You must NOT include any other code and data in a namespace package’s
__init__.py. Even though it may appear to work during development, or
when projects are installed as .egg files, it will not work when the
projects are installed using “system” packaging tools – in such cases
the __init__.py files will not be installed, let alone executed."

Tested by running a local PyPi server using
https://pypi.python.org/pypi/pypiserver and installing with this
modification to both this repo and performanceplatform-collector. You
can then successfully import performanceplatform.collector, which was
not the case before this modification.