Install requirements for SSL SNI
We use SNI to serve up *.<env>.performance.service.gov.uk in production,
sadly python2 does not support SNI out of the box which causes our
AdminAPI to fail when connecting to Stagecraft in production.

There is information about it in:
  - http://docs.python-requests.org/en/latest/community/faq/
  - https://stackoverflow.com/questions/18578439/using-requests-with-tls-doesnt-give-sni-support/18579484#18579484
  - kennethreitz/requests#749 (comment)

These point to using this combination of libraries to replace the
default SSL handling in python2.

These are only applicable for python2 and break with 3, so we only want
to install them when we are running python < 3