Allow data set POSTs to be chunked
When sending a large amount of data through to the performance platform
we need to be able to chunk the POST so that we don't time out and get
rejected. The user of the client library will be in charge of picking a
chunk size that makes sense for what they are submitting.

This doesn't wrap POSTs up in a transaction so we rely on the fact that
we have 'immutable' ids. So that if data gets written again it should
overwrite the existing data.