Ensure admin api does not gzip request data
Currently stagecraft cannot handle gzipped request which is a problem
for some of our longer dashboards. We should fix this in future,
however as a quick fix we will turn it off for the admin api as these
requests are rare and not much over the limit after which requests are
gzipped.