Convert DataSet.post data to list if 0 chunk size
Now that data can be a generator (lazy sequence), if the chunk size is
set to 0 (ie. no chunking) we must evaluate the sequence to avoid the
json serializer attempting to serialize a generator directly. This is
done by converting it to a list.

Note, this should only be applied to things that are iterators.