Requests by humans should not be retried.
All requests to backdrop are being exponentially backed off. This means
it takes a while for a 5xx response to be shown to the user.

Add a retry_on_error property to the Client base class. If set to False,
no requests of any type (GET, PUT etc) will be retried for the
instance of the Client.

The default value has been set to True so that the behaviour of the
collectors is not affected.