Change Request-Id header
Now we are on gov.uk infrastructure, it makes more sense to set the
Govuk-Request-Id header rather than our own request header.

Also remove checking for empty values of the request_id header, which
were adding no value, but add a test that the request_id function is
being called when it is set.